✨ Blog Detay Sayfası Yeniden Tasarımı - Review System & TOC İyileştirmeleri
Blog detay sayfasında kullanıcı deneyimini iyileştirmek için kapsamlı değişiklikler:

• Review/Rating system tamamen sıfırdan kodlandı (async/await, clean Alpine.js)
• Section sıralaması yeniden düzenlendi (Yazı → FAQ → HowTo → Yorumlar → Paylaş → Yazar)
• Tüm section'lara tutarlı gradient tasarım uygulandı (author card ile aynı)
• Meta card arkaplanı transparan yapıldı (sadece border)
• TOC'ye section başlıkları eklendi (FAQ, HowTo, Yorumlar, Paylaş, Yazar)
• TOC anında tepki + otomatik scroll (50ms debounce, desktop only)
• Border çizgileri kaldırıldı (temiz görünüm)
• HowTo responsive düzenlendi (2 sütun)
• Rating system user_id nullable yapıldı (guest ratings için)
• Favorite system route/service güncellemeleri

+++ Modules/ReviewSystem/app/Http/Livewire/Admin/ReviewStatisticsComponent.php
@@ -20,7 +20,7 @@ class ReviewStatisticsComponent extends Component
             'pending_reviews' => Review::where('is_approved', false)->count(),
             'approved_reviews' => Review::where('is_approved', true)->count(),
             'total_ratings' => Rating::count(),
-            'average_rating' => Rating::avg('rating_value') ? round(Rating::avg('rating_value'), 2) : 0,
+            'average_rating' => Rating::avg('rating_value') ? round((float) Rating::avg('rating_value'), 2) : 0,
             'ratings_distribution' => Rating::select('rating_value', DB::raw('count(*) as count'))
                 ->groupBy('rating_value')
                 ->orderBy('rating_value', 'desc')
